+++ app/api/utils/pusher/pusher-instance.tsx
@@ -1,5 +1,5 @@
 import Pusher from "pusher";
-import { config } from "../lib/config";
+import { config } from "../../lib/config";
 
 let pusherInstance: Pusher | null = null;
